backend/requirements.txt
aioquic-mitmproxy==0.9.21.1
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
blinker==1.7.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==40.0.2
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
Flask==3.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpq==1.2.0
hyperframe==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
kaitaistruct==0.10
ldap3==2.9.1
MarkupSafe==2.1.3
mitm==1.5.0
mitmproxy==10.1.6
mitmproxy-windows==0.4.1
mitmproxy_rs==0.4.1
msgpack==1.0.7
passlib==1.7.4
pbr==5.9.0
protobuf==4.25.1
publicsuffix2==2.20191221
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydivert==2.1.0
pylsqpack==0.3.18
pyOpenSSL==23.1.1
pyparsing==3.1.1
pyperclip==1.8.2
pytz==2023.3.post1
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
service-identity==23.1.0
sortedcontainers==2.4.0
sqlparse==0.4.4
toolbox==1.11.0
tornado==6.4
tzdata==2023.3
urwid-mitmproxy==2.1.2.1
Werkzeug==3.0.1
wsproto==1.2.0
zstandard==0.22.0
